fixes #885: renders comments below composition score
This commit displays all comments written by the grader below the
composition score. Clicking on a comment brings the student to the
line where the comment was written. The template logic is handled in
student/assignment/list_comments.html. Changes were also made to
diff.html in order to make the comments linkable with href anchors.

I don't particularly like adding the second argument to
render_comment in diff.html, but its necessary in order to make sure
that each ID occurs only once (because I used render_comment) to
render the comments twice.

Also a lot of the logic is handled in the list_comments template,
and some of that might be better rewritten in Python. I think that
its simple enough to not be a problem right now.