Improved Moss integration (#1092)
* parse moss results into database

* fix moss job; add results viewer

* check both exact and swapped duplicate

* Multi-file support; recalculated similarity

Multiple files are now concatenated into a single one to send to Moss, with the
results transformed back to per-file line numbers.

The parser now discards Moss's similarity percent and instead recalculates it
as [# of lines in Moss result but not in template / # of lines not in template]

The time the result was created is now shown in the table.

* Fixed sorted by similarity

* Adds Moss Results box to student overview

* Fix assignment moss results page

* Add moss results to student assignment page

* Hide Moss results box if none exist

* Adds submitter emails to assignment Moss page

* Fix assignment moss page length

* Remove moss results from student assignment page

* Allow searching names on assignment moss page

* add tagging support

* fix interface for moss tags

* make descriptions clearer

* start fixing bugs -- not including midfile matches

* added migration file

* Fixes and changes to how MossResults are stored

MossResults now contain primary and secondary submissions. For a run
that cares about all submissions, this will create two MossResult
objects for every result Moss sends.

MossResults also store the time the run was made at. The assignment
Moss view now only displays results from the last run for that
assignment. The student Moss view displays across runs, but filters out
runs between the same pair of submissions, preferring the most recent.

It's also possible now to specify how many results to request from Moss.
The default was 250; I'm not sure what Moss's limit is.

* remove stale migration for moss_results

* re-add that old migration

* my bad:  let's delete that migration again

* forgot to add migration file