Fix test run documentation on Windows
The pytest command referenced in the makefile fails on Windows with a
cryptic OSError: "The handle is invalid".

A full log of a test run can be found here:
https://github.com/Cal-CS-61A-Staff/ok/files/1916144/pytest-windows.log.txt

Running the tests via the management subcommand, on the other hand,
works without issues.

This commit changes the documentation to use the management subcommand
instead of the pytest command in the makefile. This might make it easier
for new contributors to get ramped up and verify their installations and
avoid confusion due to Windows and *nix differences.