Disable job and highlight tests on Windows
The rq module only supports *nix and as such the job tests fail on
Windows with an AttributeError: "module 'os' has no attribute 'fork'".

Similarly, the highlight tests depend on the *nix patch executable and
as such the tests currently fail on Windows with a FileNotFoundError.

This commit disables running of the job and highlight tests on Windows
to reduce these false positives which might make it easier for new
contributors to verify their installations.