Fix files tests on Windows
The files tests assume that the default encoding on the platform executing the
tests is UTF-8. However, on Windows CP1252 is the default encoding which leads
the file tests to fail case with a UnicodeDecodeError.

This commit changes the files tests to explicitly open the files as UTF-8 via
the io module which makes the tests pass on Windows.