Skip Redis tests if Redis is not available
Currently all tests that rely on Redis fail with a ConnectionError
unless the redis-server is run manually before starting the test runner.

This behavior could be confusing to some new contributors. As such, this
commit skips the tests that rely on Redis unless the service is actually
running.

Additionally, checking for Redis at test runtime lets us remove the
explicit ignoring of the test when running on Travis.