Add integration tests for cloud storage
The new integration tests run slightly modified versions of the existing
tests for the local storage backend. For Google Cloud Storage, the
integration tests can be enabled by setting the environment variables
`GCP_STORAGE_KEY` and `GCP_STORAGE_SECRET`. The code is designed to be
reusable across a wide range of cloud storage backends which will make
onboarding of other backends (e.g. Azure Blob Storage) easier in the
future.

This commit also fixes two issues in the storage implementation that
were uncovered by the tests:

1) The use of `datetime.strftime('%s')` relies on platform-specific
undocumented behavior and may not work across Python versions or
deployment environments. This was switched to a cross-platform way of
converting a datetime to epoch via `datetime.timestamp()`.

2) Submitting malicious files to Google Cloud Storage returned a 400
error rejecting the file. To mirror the behavior of the local storage
provider, "special" parts of files uploaded to non-local storage (such
as "." and "..") now get sanitized to underscores.

Furthermore, the commit also:

3) Updates to the latest version of apache-libcloud which improves
handling of missing files. Without this update, fetching non-existing
blobs produces a 302 result and `test_malicious_local_get_blob` fails.

4) Adds new tests to verify the results of binary downloads (previously
only text downloads were verified).

5) Disables parallel test execution on CircleCI. The cloud storage tests
rely on mutating environment variables before creation of the Flask app
which means that the tests are no longer guaranteed to be threadsafe.