Adds Owner field to EditClientForm.
Existing of owner's of OAuth clients as well as admins
can now transfer ownership by changing the owner email
for the respective OAuth client.

This change attempts to respect the current "ownerless"
state of some OAuth clients by allowing them to remain
"ownerless" even during future client setting edits.

If the owner is changed however it is verified that the
new owner does indeed exist and is a staff member of some
course. (Considering if they weren't they would be unable
to access the `/admin/client` page to manage the OAuth
client.)

Signed-off-by: Colin Schoen <cschoen@berkeley.edu>