Ensure that all settings match old values
This fixes a regression introduced in ec06b9c where the settings were
refactored into a class hierarchy but some values were changed by
accident as part of this refactor.

The fix was verified by logging out the configs on the commit before the
refactor, applying the refactor, logging out the configs again and then
diffing the results.

Snippet used to log the configs:

```py
with open('config-' + app.config['ENV'] + '.txt', 'w') as fobj:
    config = str(app.config).replace(', ', ',\n')
    fobj.write(config)
```

Paired with @liliankasem