Create custom docker image for worker
AppService for Containers seems to have an issue with the custom run
command for the Docker image. According to the logs, the worker's
AppService was starting an instance of the OK server instead of the
worker. As a simple work-around, this change introduces a Dockerfile for
the OK worker which bakes the correct CMD into the image. This does
create some level of redundancy, but on the bright side it also means
that in the future we could publish a more optimized image for the
worker, e.g. excluding the frontend assets, etc.