Remove hard-coded links to static files (#1299)
* Remove hard-coded links to static files

This change replaces all direct links to static files (such as
`/static/foo/bar`) with their equivalent `url_for` call. This introduces
consistency with code that creates links to the app's Flask routes and
also ensures that the links to the static files will always be correct
even if the application is for example hosted at a non-root path.

* Fix route not found error

The index routes are specific to various blueprints so we can't use
`url_for('.index')` in the error pages. Passing the application root as
a global context variable to the Flask templates works around this.