Generate Sendgrid API key during Azure setup (#1311)
* Increase minimum number of pods

This increases availability of the services.

* Generate Sendgrid API key during Azure setup

We've been seeing some instability in production with the runtime
creation of API keys approach. This is because of race conditions where
one thread creates an API key and then at a later time another thread
needs to re-create the key (e.g. due to the worker or pod crashing and
being restarted) thus invalidating the first one.

* Fix formatting