Run nginx and gunicorn in docker via supervisor
Currently the nginx process is run in the background of the docker
container and the gunicorn process is run in the foreground. This means
that if the nginx process crashes, it won't get restarted and the nginx
logs don't show up in the docker logs since only the foreground gunicorn
logs are captured. More generally, it is also not recommended to run
multiple processes in a single docker container unless a process manager
is used (more information can be found in [1]).

This change introduces the supervisor process manager to run as the
single foreground process in docker. The supervisor process in turn
takes care of running both nginx and gunicorn.

Note that supervisor has its own environment variable isolation and
doesn't have an easy way of passing all the environment variables to the
child process. As a work-around to avoid having to declare all
environment variables multiple times, we first dump the environment into
files and then reload them in the child process from the files via the
envdir utility [2].

[1] https://docs.docker.com/config/containers/multi-service_container/
[2] https://github.com/jezdez/envdir