Fix download and raw buttons on Firefox (#1316)
On Firefox, the `<button>` element wrapping the `<a>` link swallows the
click event which leads the download and raw buttons to not perform any
action. Note that this same behavior does not reproduce on Chrome.
Switching the `<button>` to a `<div>` fixes the behavior on Firefox
while still maintaining the same styling as before and also works on
Chrome.