Enable autopy-server to replicate across nodes
Currently the autopy-server stores the autograder files on the local
file-system via flask-admin. In Kubernetes, the file-system is managed
via a persistent volume claim. However, the volume can only attach to a
single node which means that all autopy-server pods have to be scheduled
to the same node which limits fault tolerance and scalability.

This change switches the autopy-server to use the new direct Azure Blob
Storage integration for flask-admin which stores the autograder files in
a non-local storage. As such, we can remove the persistent volume claim
and autopy-server pods can now be scheduled to arbitrary nodes.